backend/crawler/pyproject.toml
[project]
name = "chess-behavioral-crawler"
version = "0.1.0"
description = "A polite, user-centric crawler for fetching and analyzing chess games."
authors = [
    { name = "Jacob", email = "jacobfluttereddev@gmail.com" }
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-chess>=1.10.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
